+++ index.js
@@ -1,4 +1,4 @@
-const server = require('./src/server');
+const server = require('./src/server.factory');
 const minimist = require('minimist');
 const args = minimist(process.argv.slice(2));
